[TASK] Update coding styleguide to PER + additional configs
This change adapts our coding guidelines
and rules with PHP-CS-Fixer, allowing for
* PER (a.k.a. PSR-12)
* no_useless_nullsafe_operator (only use nullsafe if allowed)
* only single-spaces in arrays

This effectively adapts our code base
* use "public const" instead of "const" where visibility is not defined
* one whitespace between "declare()" and "namespace" declarations
* Use (float) instead of (double) casting
* Additional space in annonymous classes (PER)
* grouped "use" statements (classes, function, constants)
* only one space between items in array lists
* Use (bool) instead of (boolean)

Resolves: #98585
Releases: main, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76100
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/backend/Classes/Form/Element/InputTextElement.php
@@ -211,7 +211,7 @@ class InputTextElement extends AbstractFormElement
                 $itemValue = (int)$itemValue;
             } elseif (in_array('double2', $evalList, true)) {
                 $valueType = 'double';
-                $itemValue = (double)$itemValue;
+                $itemValue = (float)$itemValue;
             }
             $rangeAttributes = [
                 'id' => $id,
